Added basic shader and cleaned up workflow

+++ Xcode/VikenEngine/VEngine/VEngine/External/EBO.cpp
@@ -8,7 +8,7 @@
 
 #include "EBO.h"
 
-EBO::EBO(GLfloat *indices , GLsizeiptr size)
+EBO::EBO(GLuint *indices , GLsizeiptr size)
 {
     glGenBuffers(1, &ID);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER,ID);
